\name{LinearRegPackage-package}
\alias{LinearRegPackage-package}
\alias{LinearRegPackage}
\docType{package}
\title{
\packageTitle{LinearRegPackage}
}
\description{
\packageDescription{LinearRegPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LinearRegPackage}
\packageIndices{LinearRegPackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{LinearRegPackage}

Maintainer: \packageMaintainer{LinearRegPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
