% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg_Class.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Computes Multiple Linear Regression For Given Formula And Data}
\description{
Linreg class calculates the computations required for multiple regression model
    using the ordinary least square method. It also provides methods like print, plot,
    resid, coef and summary.
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{est_beta}}{vector.}

\item{\code{y_pred}}{matrix.}

\item{\code{resid_e}}{matrix.}

\item{\code{deg_freed}}{numeric.}

\item{\code{resid_var_e}}{matrix.}

\item{\code{var_est_beta}}{vector.}

\item{\code{t_val_beta}}{matrix.}

\item{\code{data_set}}{character.}
}}


