% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinReg_Class.R
\docType{class}
\name{linregClass-class}
\alias{linregClass-class}
\alias{linregClass}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{est_beta}}{vector.}

\item{\code{y_pred}}{matrix.}

\item{\code{resid_e}}{matrix.}

\item{\code{deg_freed}}{numeric.}

\item{\code{resid_var_e}}{matrix.}

\item{\code{var_est_beta}}{vector.}

\item{\code{t_val_beta}}{matrix.}
}}


