% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge_Reg_Class.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge Regression}
\description{
Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{lambda}}{numeric.}

\item{\code{beta_ridge}}{vector.}

\item{\code{X}}{matrix.}

\item{\code{y_pred}}{matrix.}

\item{\code{data_set}}{character.}
}}


