% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualizing Mean Delay Of Flights For Different Airports In USA}
\usage{
visualize_airport_delays()
}
\value{
A girafe object of an interactive plot showing mean delay of each airport
}
\description{
We visualize the mean delay of flights for different airports
    by plotting the Longitude vs Latitude respectively. Upon hovering over
    the points in the plot, each airport's mean delay with faa code can be
    seen.
}
