% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge_Reg_Class.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Computes Ridge Regression For Given Formula And Data}
\description{
ridgereg class calculates the computations required for multiple 
    regression model for ridge regression. It also provides methods like predict, 
    plot, coef and show.
}
\details{
Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{Formula for the ridge regression.}

\item{\code{data}}{Data frame containing the dependent and independent variables.}

\item{\code{lambda}}{Number for the lambda value.}

\item{\code{beta_ridge}}{The vector for the coefficients.}

\item{\code{X}}{The model matrix.}

\item{\code{y_pred}}{Matrix containg the predicted values.}

\item{\code{data_set}}{character.}
}}


